restricts deleting item to users with appropriate permissions

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -40,9 +40,13 @@ const Mutations = {
     async deleteItem(parent, args, ctx, info) {
         const where = { id: args.id};
         // find the item
-        const item = await  ctx.db.query.item({ where} , `{id, title}`);
+        const item = await  ctx.db.query.item({ where} , `{id title user { id }}`);
         // check if they have the permissions
-        //TODO
+        const ownsItem = item.user.id === ctx.request.userId;
+        const hasPermissions = ctx.request.user.permissions.some(permission =>['ADMIN', 'ITEMDELETE'].includes(permission));
+        if (!ownsItem && !hasPermissions) {
+          throw new Error("You don't have the right permissions to do that!");
+        }    
         // delete it
         return ctx.db.mutation.deleteItem({where}, info);
       },
